Continued rearrangement and made some changes in MCCE code.
Got to line ~200. Now I need to check the datatype of
"new_predicted_data" in earlier commits. I am wondering if the datatype
of this data, and the datatypes of the columns, where the same as they
are now, and if this is a problem for the algorithm I implemented in
later lines of code.