Made some changes to Gower (normalization) and started
first sim on Markov with randomForest. Also checked trees and the
distributions od D_h. The distributions of D_h are good for all
categorical variables for the binarized data, while they are worse for
the numerical variables (since they have outliers I think). Moreover, it
seems like the distributions that are generated for the categorical data
is not good.