Added change to generation algo sampling from trees.
It has been tested locally and is being tested on Markov now! Looking
rather good. These results will be important in deciding whether the
trees for the categorical data are bad or if there was simply a big
problem with the tree node sampling in the categories (for more than two
categories). Exciting!!