Made many changes in this branch  (#1)
* Code for rforest with ranger.

* Implemented ranger as pred model. Checked trees for modelling distr.

Looks good! Also, I removed all "?" from the dataset (treating them as
NAs)

* Made some changes to Gower (normalization) and started

first sim on Markov with randomForest. Also checked trees and the
distributions od D_h. The distributions of D_h are good for all
categorical variables for the binarized data, while they are worse for
the numerical variables (since they have outliers I think). Moreover, it
seems like the distributions that are generated for the categorical data
are not good.

* Looked at the numerical covariates --> not much to do!

* Updated the sampling from tree end node in generation algorithm.

This needs to be tested before proceeding!!!

* Added change to generation algo sampling from trees.

It has been tested locally and is being tested on Markov now! Looking
rather good. These results will be important in deciding whether the
trees for the categorical data are bad or if there was simply a big
problem with the tree node sampling in the categories (for more than two
categories). Exciting!!

* Figured out why ANN "seemed" so bad all along!!!!!

Also added some results. Now only missing to make D_h distribution
"good" for numerical variables and fixing calculation of Gower!

* Made some more changes.

* Made some changes to normalization factor in Gower!

* Made some more changes to Gower (division by 0 error)

Perhaps it is a good idea to redefine this in the theory also!

* Changed the normalization factor in Gower to M-m

Co-authored-by: alexaoh <alexaoh@users.noreply.github.com>