Merge original one-hot encoding. (#2)
IT IS POSSIBLE THAT THIS IS STILL A PROBLEM: I am having a problem with predicting from D_h in the post-processing,
since I need to make the model matrix again for ANN. This is not always
possible (e.g. when D_h is missing levels in some covariates). How else
could I solve this?? I need help! 

Co-authored-by: alexaoh <alexaoh@users.noreply.github.com>