Worked on VAE --> Got some "good" results! Next:
Next is post-processing the samples from VAE! I tried to do this quickly
in the MCCEalgo1.R-file, but I had problems with interchanging data
sets. Try again (more calmly) tomorrow and see if I am able to add the
actionable criterion also! (which is not needed for the trees in MCCE,
but is needed in ModMCCE, since we are not conditioning on anything in
the VAE)